Basically what it does is filter the data from nyc open data about vehicle collision reports and displays a histogram showing the total numb
Basically what it does is filter the data from nyc open data about
vehicle collision reports and displays a histogram showing the total
number of injuries by borough, borough with the most reported incidents,
most common cause, etc. When you filter by either date, address,
borough, or zip code it will display the ten reports for each page and
plot the points in which the collision occured on the map.